// src/types/customerTypes.ts
// Define a "forma" da entidade Cliente (Customer), com base no JSON da API.
export type Customer = {
    id: number;
    name: string;
    nickname: string | null;
    cpf: string | null;
    rg: string | null;
    bornDate: string | null;
    gender: string | null;
    fatherName: string | null;
    motherName: string | null;
    spouse: string | null;
    email: string | null;
    company: string | null;
    occupation: string | null;
    street: string | null;
    number: string | null;
    neighborhood: string | null;
    city: string | null;
    uf: string | null;
    cep: string | null;
    complement: string | null;
    isBlacklisted: boolean | null;
    obs: string | null;
    phone01: string | null;
    phone02: string | null;
    phone03: string | null;
    reference01: string | null;
    reference02: string | null;
    reference03: string | null;
    isActive: boolean;
    createdAt: string;
    updatedAt: string;
};